refactor(orders): wave-3 slice 4 work-order application scenario tables (#3949)
Migrate the five orders_application_work_order_application_part* shards to a thin family runner with support scenario/expectation tables, preserving descriptions for the description-baseline gate.
